Cleanup WebSite with StyleCop
I cleaned up the code of the Website with Resharper and StyleCop. I
haven't touched the account stuff, since it isn't really SourceBrowser
code.
Install Resharper and StyleCop on your machines too, this really helps to
provide good code quality. With those two it is as easy as pressing
"Alt-Enter" to fix a quality issue.