Removed /Hacks and added Roslyn DLLs to website.
In the past we had a bunch of issues with this approach, but it seems
they've fixed the bugs we were running into.

Our /Hacks approach didn't work for VB and no longer worked for C#. Now
there are a bunch of new .Desktop DLLs we would have needed to include
as well.