Displays "await" page when repo is being processed
The web server is still processing our request in the background, even
if the user navigates to another page.
The "await" page is visible to users who browse while another user
uploads, or to the user who started the upload and decided to browse
that repo. In the future, this page should use SignalR to communicate
with the server and show live updates of the processing (including
refreshing at the end)