Update existing repo feature:
when a user submits a repository that already exists, a page is returned where he can choose if he wants to go to the existing source or force the update.

Added a new action "Update" (and a View) in the UploadController.
Refactored some of the ProcessSolution logic into the UploadRepository class.
As suggested in the comments (by Amadeus Wieczorek), added support for multi-solution repository and parallel solution processing.